pyproject.toml
[tool.poetry]
name = "poc-caching-ddd-graphql"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.3"
strawberry-graphql = {extras = ["opentelemetry"], version = "^0.41.0"}
django-extensions = "^3.0.9"
Werkzeug = "^1.0.1"
aioredis = "^1.3.1"
dacite = "^1.5.1"
django-cors-headers = "^3.5.0"
protobuf = "^3.13.0"
factory-boy = "^3.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
mypy = "^0.790"
black = {version = "^20.8b1", allow-prereleases = true}
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
